fix effort foreign key

+++ db/migrate/20190825193054_create_segments.rb
@@ -3,7 +3,7 @@ class CreateSegments < ActiveRecord::Migration
     create_table :segments do |t|
       t.integer :points
       t.string :name
-      t.integer :strava_id
+      t.bigint :strava_id
       t.float :max_grade
       t.float :average_grade
       t.integer :climb_category
